+++ tests/test_model_zoo.py
@@ -7,6 +7,12 @@ import torchaudio
 from torchaudio import models as audio_mdl
 from torchvision import models as vision_mdl
 
+from tests.shifted_window_attention_patch import (
+    ExportableShiftedWindowAttention,
+    ExportableSwinTransformerBlock,
+)
+from torch_to_nnef.tract import tract_version_greater_than
+
 from .utils import (  # noqa: E402
     _test_check_model_io,
     remove_weight_norm,
@@ -113,10 +119,30 @@ INPUT_AND_MODELS += [
     ),
 ]
 
-# export pretrained work but multi_head given slightly different values
+# export pretrained work but multi_head giving slightly different values
 INPUT_AND_MODELS += [
     (torch.rand(1, 3, 224, 224), vision_mdl.vit_b_16(pretrained=False)),
 ]
+
+
+# swin_transformer {
+# need slice with stride
+if hasattr(vision_mdl, "swin_transformer") and tract_version_greater_than(
+    "0.19.0"
+):
+    vision_mdl.swin_transformer.ShiftedWindowAttention = (
+        ExportableShiftedWindowAttention
+    )
+    vision_mdl.swin_transformer.SwinTransformerBlock = (
+        ExportableSwinTransformerBlock
+    )
+    data = torch.rand(1, 3, 224, 224)
+    mdl = vision_mdl.swin_t()  # pretrained=False
+    mdl.eval()
+    mdl(data)  # precompute attn mask and few shapes
+    INPUT_AND_MODELS += [(data, mdl)]
+
+# }
 
 
 @pytest.mark.parametrize("test_input,model", INPUT_AND_MODELS)