+++ torch_to_nnef/op/quantized.py
@@ -13,7 +13,7 @@ from nnef_tools.model import Tensor as NTensor
 
 from torch_to_nnef.exceptions import TorchToNNEFNotImplementedError
 from torch_to_nnef.op.primitive import _add_single_output_op
-from torch_to_nnef.tract import tract_version_lower_or
+from torch_to_nnef.tract import tract_version_lower_than
 
 
 def _torch_qtensor_to_ntensor(g, tensor, name):
@@ -178,7 +178,7 @@ def _conv(
     # apply expansion to align inputs with weight {
     for _ in range(input_node.rank - len(conv_weight.shape)):
         conv_weight = conv_weight.unsqueeze(0)
-    if conv_bias is not None and tract_version_lower_or("0.18.1", False):
+    if conv_bias is not None and tract_version_lower_than("0.18.1"):
         for _ in range(input_node.rank - len(conv_bias.shape)):
             conv_bias = conv_bias.unsqueeze(0)
     # }