+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 ## Unreleased
+### Added
+- `roll`, `new_zeros`, `zeros` operators
+- `pow` operator now support negative and scalars as exponent
+
+### Fix
+- `rsub` & `remainder` operator with constant should be precomputed output constants
 
 ## [0.6.10] - 2022-11-07